Fill in property set identifiers when updating.
This is a bit of a cludge, the frontend keeps property set identifiers
the same when updating them, as it knows the indicies as they are being
written.

We can't really do that here, but we do have some chance. If the property
set name stays the same, then we can aling the identifier by comparing
the old values. This should cover most use cases, but one can also use
an explicit id as well.