Improve handling of CSV source fields.
We're hit by [Terraform SDK Issue 258](https://github.com/hashicorp/terraform-plugin-sdk/issues/258)
here, which causes the elements of the nested structure to be
filled in even if it's not set in the record. This is a very
unfortunate part of the SDK.

Becase the map contains the empty string unconditionally and
we can't use GetOk like we could on the top level, we have to
check if the string is empty instead.

For boolean fields, we can't even do this properly, as it's
always false or true, so our optional elements in the Schema
will always be set, fortunately the Spark options will still
do the right thing here.