Improve terraform validation.
The types on the backend are shared for table and feature checks,
and have the same API. But, the types are a bit more restrictive
for features because they must be able to be computed in a single
pass during feature generation, and can't require a new plan.

So we don't allow one to _input_ a unique constraint, but we will
show it if it's there (it will result in a skipped check if the
feature generator runs it).

I tried using MaxItems 0 for this, but because go it very weak,
it confused 0 with empty and just ignored it. So I instead have
it conflict with its base, so you can't input it.